requirements.txt
crewai
crewai-tools
langchain-google-genai
google-generativeai
requests
python-dotenv
Pillow
